added debug data and added button --> read description
Added two testfeilds to show the look of the applcations shown after
searching( only for debug and will be removed in the future).

And also changed the design of the applcation list so it has a button
("View") that later on will make it possible to view the whole
application that is connected to the button. (all the applications will
have its own button. just so we are clear).